+++ src/common/helpers/logging/logger.js
@@ -1,6 +1,6 @@
 import { pino } from 'pino'
 
-import { loggerOptions } from './logger-options.js'
+import { loggerOptions } from '../../../plugins/logger-options.js'
 
 const logger = pino(loggerOptions)
