+++ src/plugins/pulse.js
@@ -1,5 +1,5 @@
 import hapiPulse from 'hapi-pulse'
-import { createLogger } from './logging/logger.js'
+import { createLogger } from '#src/common/helpers/logging/logger.js'
 
 const tenSeconds = 10 * 1000
